% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{afli_afli}
\alias{afli_afli}
\title{afli.afli}
\usage{
afli_afli(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu afli í
afli gagnagrunninum.
}

