% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{afli_lineha}
\alias{afli_lineha}
\title{afli.lineha}
\usage{
afli_lineha(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu linea í
afli gagnagrunninum.
}
