% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{afli_sjalfvirkir_maelar}
\alias{afli_sjalfvirkir_maelar}
\title{afli.sjálfvirkir_maelar}
\usage{
afli_sjalfvirkir_maelar(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu sjálfvirkir_maelar í
afli gagnagrunninum.
}

