% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{afli_stofn}
\alias{afli_stofn}
\title{afli.stofn}
\usage{
afli_stofn(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu stofn í
afli gagnagrunninum.
}

