% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbooks.R
\name{afli_toga}
\alias{afli_toga}
\title{afli.toga}
\usage{
afli_toga(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu toga í
afli gagnagrunninum.
}

