% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata.R
\name{ata_hafsyni}
\alias{ata_hafsyni}
\alias{ata_thurrvigt_tegund}
\alias{ata_rummal_thattur}
\alias{ata_thurrvigt_thattur}
\alias{ata_talningamadur}
\alias{ata_astand}
\alias{ata_votvigt}
\alias{ata_verkefni}
\alias{ata_thurrvigt}
\alias{ata_tegundir_hopar}
\alias{ata_talning}
\alias{ata_stod}
\alias{ata_stig}
\alias{ata_sofnunartaeki}
\alias{ata_snid}
\alias{ata_shortcut}
\alias{ata_rummal}
\alias{ata_prosomlengd}
\alias{ata_lengdir}
\alias{ata_leidangur}
\alias{ata_kynthroski}
\alias{ata_hopar_stig}
\alias{ata_hopar}
\alias{ata_flurljomun}
\alias{ata_eggjaframleidsla}
\title{Ástusýni}
\arguments{
\item{con}{tenging við mar}
}
\description{
Ástusýni
}
