% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_brad_lengdir}
\alias{bio_brad_lengdir}
\title{Fæða lengdir}
\usage{
bio_brad_lengdir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu f_lengdir
í biota gagnagrunninum.
}
