% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_brad_tyngdir}
\alias{bio_brad_tyngdir}
\title{Fæða þyngdir}
\usage{
bio_brad_tyngdir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu f_hopar
í biota gagnagrunninum.
}
