% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_faedutegundir}
\alias{bio_faedutegundir}
\title{Tegundir fæðu}
\usage{
bio_faedutegundir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\value{
SQL fyrirspurn
}
\description{
Fallid myndar tengingu við "view" toflu f_tegundir
í biota gagnagrunninum.
}
