% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_kvarnir}
\alias{bio_kvarnir}
\title{Kvarnir}
\usage{
bio_kvarnir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við "view" toflu kvarnir
í biota gagnagrunninum.
}
