% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_lengdir}
\alias{bio_lengdir}
\title{Lengdarmaelingar}
\usage{
bio_lengdir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við "view" toflu lengdir
í biota gagnagrunninum.
}
