% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_linir}
\alias{bio_linir}
\title{Linir}
\usage{
bio_linir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu linir
í biota gagnagrunninum.
}
