% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_numer}
\alias{bio_numer}
\title{Talningar}
\usage{
bio_numer(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við "view" toflu numer
í biota gagnagrunninum.
}
