% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_ranfiskar}
\alias{bio_ranfiskar}
\title{Fæða ránfiskar}
\usage{
bio_ranfiskar(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu f_fiskar
, f_flokkar og f_lenfl í biota gagnagrunninum.
}
