% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_skala_med_toglengd}
\alias{bio_skala_med_toglengd}
\title{Skala með toglengd}
\usage{
bio_skala_med_toglengd(
  st_len,
  min_towlength = 2,
  max_towlength = 8,
  std_towlength = 4
)
}
\arguments{
\item{lengdir}{fyrirspurn á biota.lengdir}
}
\value{
fyrirspurn með sköluðum fjölda í lengdarbili
}
\description{
Þetta fall skala lengdardreifingar með toglengd úr ralli
}
