% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_skala_med_toldum}
\alias{bio_skala_med_toldum}
\title{Skala með töldum}
\usage{
bio_skala_med_toldum(lengdir)
}
\arguments{
\item{lengdir}{fyrirspurn á biota.lengdir}
}
\value{
fyrirspurn með sköluðum fjölda í lengdarbili
}
\description{
Þetta fall skala lengdardreifingar með töldum fiskum úr ralli
}
