% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_skjoda_lengdir}
\alias{bio_skjoda_lengdir}
\title{Skjóða lengdir}
\usage{
bio_skjoda_lengdir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu skjoda_lengdir
í biota gagnagrunninum.
}
