% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_skjoda_numer}
\alias{bio_skjoda_numer}
\title{Skjóða númer}
\usage{
bio_skjoda_numer(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu skjoda_numer
í biota gagnagrunninum.
}
