% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_snikjudyr}
\alias{bio_snikjudyr}
\title{Sníkjudýr}
\usage{
bio_snikjudyr(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu snikjudyr
í biota gagnagrunninum.
}
