% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{bio_stodvar}
\alias{bio_stodvar}
\title{Stodvarupplysingar}
\usage{
bio_stodvar(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við "view" toflu stodvar_mar_v
í channel gagnagrunninum.
}
