% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biota.R
\name{bio_valkvarnir}
\alias{bio_valkvarnir}
\title{Valkvarnir}
\usage{
bio_valkvarnir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Fallid myndar tengingu við "view" toflu valkvarnir
í biota gagnagrunninum.
}
