% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate  1H-NMR spectra}
\usage{
calibrate(X, f1, f2)
}
\arguments{
\item{X}{A matrix of spectra containing ppm values in the first column
and intensity values for each sample in the remaining columns. If a
matrix with samples x ppm is provided instead, it will be transposed
prior to plotting.}

\item{f1}{Chemical shift of the leftmost ppm value to plot containing
the spectral reference (e.g. ERETIC, TSP, DSS, glucose...).}

\item{f2}{Chemical shift of the rightmost ppm value to plot containing
the spectral reference (e.g. ERETIC, TSP, DSS, glucose...).}
}
\value{
A matrix of 1H-NMR spectra with the calibrated ppm scale.
}
\description{
Function to align spectra according to a particular signal.
}
\examples{
data("urine")
cs <- calibrate(urine, 0.1, -0.1)

}
\author{
Goncalo Graca, Imperial College London
}
