% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_fjolst.R
\name{cn_fjolst}
\alias{cn_fjolst}
\title{Conform to fjolst column heading}
\usage{
cn_fjolst(d)
}
\arguments{
\item{d}{}
}
\value{
dataframe
}
\description{
The function converts column of a dataframe, returning "."
where there is an "_".
}
