% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_mar.R
\name{cn_mar}
\alias{cn_mar}
\title{Conform to mar column heading}
\usage{
cn_mar(d)
}
\arguments{
\item{d}{}
}
\value{
dataframe
}
\description{
The function converts column of a dataframe, returning "_"
where there is a ".".
}
