% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.R
\name{encode_zchords}
\alias{encode_zchords}
\title{Encode zchords}
\usage{
encode_zchords(
  data,
  col.names = c("lat", "lon"),
  dx = 1,
  dy = 0.5,
  invalids = TRUE
)
}
\arguments{
\item{data}{database query}

\item{col.names}{column names containing lat and lon coordinates}

\item{dx}{rounding in the x direction}

\item{dy}{rounding in the y direction}

\item{invalids}{logical}
}
\value{
database query with aggregated box
}
\description{
Encode zchords
}
