% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magi.R
\name{faeda}
\alias{faeda}
\alias{faeda_tegundir}
\alias{faeda_ranfiskar}
\alias{faeda_thyngdir}
\alias{faeda_lengdir}
\title{Magasyni Hafrannsóknarstofnunar}
\usage{
faeda_tegundir(con)

faeda_ranfiskar(con)

faeda_thyngdir(con)

faeda_lengdir(con)
}
\arguments{
\item{con}{Tenging við Oracle}
}
\value{
SQL fyrirspurn
}
\description{
Magasyni Hafrannsóknarstofnunar
}
