% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{fiskifelag_oslaegt}
\alias{fiskifelag_oslaegt}
\title{Gamall afli (fyrir 94)}
\arguments{
\item{mar}{tenging við mar}
}
\value{
aflatölur eftir höfn, skipi, veiðarfærði, árum og mánuðum
}
\description{
Afli skráður af Fiskifélaginu
}
