% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genemaps.R
\name{genemaps}
\alias{genemaps}
\title{Title}
\usage{
genemaps(
  lonlat,
  genotype,
  het2hom = TRUE,
  sample = NULL,
  genoinfo = NULL,
  mapres = NULL,
  mapcrs = "+proj=longlat +datum=WGS84"
)
}
\arguments{
\item{mapcrs}{}
}
\value{

}
\description{
Title
}
