% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomaps.R
\name{genomaps}
\alias{genomaps}
\title{Create a genomaps object}
\usage{
genomaps(geno, maps)
}
\arguments{
\item{geno}{A margeno object.}

\item{maps}{A marmaps object.}
}
\value{
A genomaps object.
}
\description{
The genomaps object is the central data structure for the mar package. It combines the genetic data
from the margeno object and the spatial information from the marmaps object, providing a
comprehensive representation of the genetic and geographic data, connected by the sample IDs.
}
\examples{
sample_id <- c("sample1", "sample2", "sample3")
genotype <- matrix(c(1, 2, 0, 0, 1, 1, 2, 0, 1, 2, 2, 1), nrow = 4, ncol = 3)
ploidy <- 2
geno <- margeno(genotype, ploidy, sample.id = sample_id)

lonlatdf <- data.frame(
    id = sample_id,
    longitude = c(-122.1, -121.9, -122.0),
    latitude = c(37.8, 37.7, 37.9),
    stringsAsFactors = FALSE
)
maps <- marmaps(lonlatdf, mapres = NULL, mapcrs = "EPSG:4326")

genomaps(geno, maps)
}
