% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lttle_helpers.R
\name{geo_convert}
\alias{geo_convert}
\title{Converts degree-minutes-seconds to degree decimals}
\usage{
geo_convert(x)
}
\arguments{
\item{x}{Vector of the form DDMMSS (degree-minutes-seconds)}
}
\description{
This is the same function as \code{geo::geoconvert.1} , except function does not
return a printout of "error > 60 min"
}
\author{
Hoskuldur Bjornsson <hoski@hafro.is>
}

