% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv.R
\name{get_skyndilokun}
\alias{get_skyndilokun}
\title{Get skyndilokun

Gets information about a particlar quick-area closure number
for a given year.}
\usage{
get_skyndilokun(con, year, nr)
}
\arguments{
\item{con}{connection to Orcacle}

\item{year}{year of closure}

\item{nr}{the closure number}
}
\description{
Get skyndilokun

Gets information about a particlar quick-area closure number
for a given year.
}
