% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gmexp}
\alias{gmexp}
\title{gmexp}
\format{
\subsection{\code{gmexp}}{

An S3 \code{genomaps} object created using \code{\link[=genomaps]{genomaps()}} function.
\describe{
\item{geno}{An S3 \code{margeno} object. See \code{\link[=margeno]{margeno()}} function.}
\item{maps}{An S3 \code{marmaps} object. See \code{\link[=marmaps]{marmaps()}} function.}
}
}
}
\source{
data-raw/gmexp_simulate.py
}
\usage{
gmexp
}
\description{
Example genomaps object created from simulated expanding population
}
\keyword{datasets}
