% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hn.R
\name{Hn}
\alias{Hn}
\title{Harmonic number}
\usage{
Hn(n)
}
\arguments{
\item{n}{"non-negative numeric" Non integer is also allowed.}
}
\value{
"numeric"
}
\description{
Harmonic number
}
\note{
Created: Wed May 30 10:51:32 CEST 2012;
      Current: Wed May 30 10:51:37 CEST 2012.
}
\author{
Feng Li, Department of Statistics, Stockholm University, Sweden.
}
