% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvalir.R
\name{hvalir_hvalir}
\alias{hvalir_hvalir}
\alias{hvalir_eistu}
\alias{hvalir_eggjastokkar}
\alias{hvalir_kynthroski}
\alias{hvalir_magasyni}
\alias{hvalir_magagreining}
\alias{hvalir_faedumaeling}
\title{Hvalir}
\usage{
hvalir_hvalir(mar)

hvalir_eistu(mar)

hvalir_eggjastokkar(mar)

hvalir_kynthroski(mar)

hvalir_magasyni(mar)

hvalir_magagreining(mar)

hvalir_faedumaeling(mar)
}
\arguments{
\item{mar}{Tenging við Oracle}
}
\value{
SQL fyrirspurn
}
\description{
Hvalir
}
