% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{hydro}
\alias{hydro}
\alias{hydro_hitamaelar}
\alias{hydro_stadir}
\alias{hydro_stodvanofn}
\alias{hydro_observation}
\alias{hydro_trolltog}
\alias{hydro_station}
\alias{hydro_sonda}
\title{Umhverfismælingar Hafrannsóknastofnunnar}
\usage{
hydro_hitamaelar(con)

hydro_stadir(con)

hydro_stodvanofn(con)

hydro_observation(con)

hydro_trolltog(con)

hydro_station(con)

hydro_sonda(con)
}
\arguments{
\item{con}{Tegning við Oracle grunn}
}
\value{
SQL fyrirspurn
}
\description{
Umhverfismælingar Hafrannsóknastofnunnar
}
