% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{kvoti_stada_summarised}
\alias{kvoti_stada_summarised}
\title{Kvóti staða summary}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\description{
Á síðari árum er þetta nokk svona:
kvoti = varanlegt + jofnsj + m_ara
stada = kvoti - afli
eftir = stada - tilf
eftir + upptaka = n_ar + onotad
}
