% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lods.R
\name{landadur_afli}
\alias{landadur_afli}
\title{Landadur afli}
\usage{
landadur_afli(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\description{
Landaður afli. Dálkurinn flokkur vísar til skipaflokks,
ef neikvæður þá er um erlent skip að ræða. Erlend skip eru
einnig þegar þetta er skrifað (1. apríl 2016) með númer frá 3815:4999.
}

