% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb.R
\name{lb_mobile}
\alias{lb_mobile}
\title{lb_base}
\usage{
lb_mobile(con, standardize = TRUE)
}
\arguments{
\item{con}{Oracle connection}

\item{standardize}{Boolean (default TRUE) if only certain variables returned}
}
\value{
A sql tibble
}
\description{
lb_base
}
