% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiskar.R
\name{lesa_kvarnir}
\alias{lesa_kvarnir}
\title{Kvarnir}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu kvarnir
í fiskar gagnagrunninum.
}
