% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiskar.R
\name{lesa_lengdir}
\alias{lesa_lengdir}
\title{Lengdarmaelingar}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu lengdir
í fiskar gagnagrunninum.
}
