% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiskar.R
\name{lesa_stodvar}
\alias{lesa_stodvar}
\title{Stodvarupplysingar}
\usage{
lesa_stodvar(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við "view" toflu v_stodvar
í fiskar gagnagrunninum.
}
