% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{lesa_synaflokka}
\alias{lesa_synaflokka}
\title{lesa synaflokka}
\usage{
lesa_synaflokka(mar)
}
\arguments{
\item{mar}{db connection}
}
\value{
db query object
}
\description{
lesa synaflokka
}
\examples{
lesa_synaflokka(mar)
}

