% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms.R
\name{lesa_vms}
\alias{lesa_vms}
\title{Lesa vms gögn}
\usage{
lesa_vms(con, year)
}
\arguments{
\item{con}{XXX}

\item{year}{XXX}
}
\description{
Athugið aðgangur er takmarkaður við fáa notendur á Hafrannsóknastofnun
}
