% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARsadsfs.R
\name{ll_sfs}
\alias{ll_sfs}
\title{Calculate Log-Likelihood Between Model and Data SFS}
\usage{
ll_sfs(model, data, missing_model_cutoff = 1e-06)
}
\arguments{
\item{model}{An object of class "sfs" containing model predictions}

\item{data}{An object of class "sfs" containing observed data}

\item{missing_model_cutoff}{Threshold for warning about SFS entries that are zeros in the model but non-zero in the data. Default 1e-6.}
}
\value{
Log-likelihood value comparing model to data
}
\description{
Calculate Log-Likelihood Between Model and Data SFS
}
\examples{
# Calculate log-likelihood between model and data SFS
model_sfs <- expsfs(lenAC=1000, N=50, ploidy=2)
data_sfs <- sfs(AC=c(1,1,0,2,0,1,1,0,0,2,30), N=50, ploidy=2)
ll <- ll_sfs(model_sfs, data_sfs)
}
