% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar-package.R
\docType{package}
\name{mar-package}
\alias{mar}
\alias{mar-package}
\title{mar: Fitting Mutations-Area Relationships using population genomic samples of a species}
\description{
Fitting Mutations-Area Relationships using population genomic samples of a species.
}
\keyword{internal}
