% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_info.R
\name{mar_views}
\alias{mar_views}
\title{mar_tables}
\usage{
mar_views(con, schema)
}
\arguments{
\item{con}{src_oracle tenging við oracle}

\item{schema}{character vector specifying schema name, e.g. "fiskar"}
}
\value{
dataframe
}
\description{
Fallið myndar tengingu við sys.all_views
}
