% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARcalc.R
\name{MARcalc}
\alias{MARcalc}
\title{Calculate the mutations-area relationship (MAR) using the power-law model}
\usage{
MARcalc(mardf, Mtype = .Mtype, Atype = .Atype)
}
\arguments{
\item{mardf}{A data frame with columns of diversity metric and area.
Output from \link{MARsampling}, \link{MARextinction} or \link{MARtheory}.}

\item{Mtype}{The diversity metric to use. Default is "M" (richness), allowed values are M, E, thetaw, thetapi.}

\item{Atype}{The area metric to use. Default is "A" (area), allowed values are A, Asq, N.}
}
\value{
A fitted power model object from sars::sar_power
}
\description{
This function calculates the MAR using the power-law model from the provided data frame.
The data frame should contain columns for the diversity metric (e.g. richness, evenness)
and the area metric (e.g. total area, area squared).
}
\examples{
area <- 1:10
mutations <- c(50, 75, 100, 125, 150, 175, 200, 225, 250, 275)
data <- data.frame(A = area, M = mutations)
mar <- MARcalc(data, Mtype = "M", Atype = "A")
summary(mar)
}
