% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARPIPELINE.R
\name{MARPIPELINE}
\alias{MARPIPELINE}
\title{Main pipeline function. Run MAR (mutations-area relationship) Analysis Pipeline}
\usage{
MARPIPELINE(
  name,
  workdir,
  genofile,
  lonlatfile,
  extra_file = list(samplefile = NULL, posfile = NULL, subsample = NULL, subvariant =
    NULL),
  filetype = c("text", "vcf", "plink"),
  option_geno = list(ploidy = 2, maxsnps = 1e+06),
  option_map = list(mapres = NULL, mapcrs = "OGC:CRS84"),
  option_sadsfs = list(sad_models = .sad_models, folded = TRUE),
  option_marext = list(scheme = .MARsampling_schemes, nrep = 10, xfrac = 0.01, quorum =
    TRUE, animate = FALSE, myseed = NULL),
  marsteps = c("data", "gm", "sfs", "mar", "ext", "plot"),
  saveobj = FALSE
)
}
\arguments{
\item{name}{Required.Character string. Base name for output files.}

\item{workdir}{Required. Character string. Working directory path where outputs will be saved. Need to be created before running the pipeline.}

\item{genofile}{Required. Character string. Path to input genomic data file. Full path to the file is required.}

\item{lonlatfile}{Character string. Path to file containing longitude/latitude coordinates. Full path to the file is required.}

\item{extra_file}{List containing optional input files:
\itemize{
\item samplefile: Path to sample metadata file
\item posfile: Path to variant position file
\item subsample: Path to file with sample IDs to subset
\item subvariant: Path to file with variant IDs to subset
}}

\item{filetype}{Character string. Input genomic file format: 'text', 'vcf', or 'plink'.}

\item{option_geno}{List of genomic data options:
\itemize{
\item ploidy: Ploidy level of samples (default: 2)
\item maxsnps: Maximum number of SNPs to analyze (default: 1000000)
}}

\item{option_map}{List of mapping options:
\itemize{
\item mapres: Map resolution
\item mapcrs: Coordinate reference system (default: WGS84)
}}

\item{option_sadsfs}{List of SAD/SFS analysis options:
\itemize{
\item sad_models: SAD models to fit
}}

\item{option_marext}{List of MAR/extinction analysis options:
\itemize{
\item scheme: Sampling schemes to use
\item nrep: Number of replicates
\item quorum: Logical; use quorum sampling
\item animate: Logical; create animation
\item myseed: Random seed for reproducibility
}}

\item{marsteps}{Character vector specifying which analysis steps to run:
"data", "gm", "sfs", "mar", "ext", "plot"}

\item{saveobj}{Logical. Whether to save intermediate objects as .rda files.}
}
\value{
Returns invisibly. Creates output files in working directory:
\itemize{
\item PDF plots for selected analysis steps
\item .rda files with analysis objects if saveobj=TRUE
}
}
\description{
MARPIPELINE performs a complete MAR analysis workflow, including data loading,
building \code{genomaps} object, Species Abundance Distribution (SAD) fitting, Site Frequency
Spectrum (SFS) calculation, sampling of geographical areas to build MAR, and extinction simulations given MAR predictions.
}
\examples{
\dontrun{
MARPIPELINE(
  name = "myanalysis",
  workdir = "path/to/output",
  genofile = "path/to/genotype.vcf",
  lonlatfile = "path/to/coordinates.txt",
  filetype = "vcf",
  marsteps = c("data", "gm", "sfs", "mar", "plot")
)
}

}
