% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARsadsfs.R
\name{MARsad}
\alias{MARsad}
\title{Fit Species Abundance Distribution (SAD) Models to Genotype Data}
\usage{
MARsad(gm, sad_models = .sad_models, predict = TRUE, folded = TRUE)
}
\arguments{
\item{gm}{A genotype matrix object containing genetic data}

\item{sad_models}{Vector of SAD model names to fit. Default uses internal mar:::.sad_models}

\item{predict}{Logical, whether to predict SFS from fitted SAD models. Default TRUE}

\item{folded}{Logical, whether to fold the SFS. Default TRUE}
}
\value{
A list with S3 class "marsad" containing:
\item{sadms}{List of fitted SAD models}
\item{AICtabs}{AIC table comparing model fits}
\item{sadsfss}{Predicted site frequency spectra if predict=TRUE}
}
\description{
Fit Species Abundance Distribution (SAD) Models to Genotype Data
}
\examples{
# Fit SAD models to genotype data in gm1001g
sad_fit <- MARsad(gm1001g)
}
