\name{pinkham}
\alias{pinkham}
\docType{data}
\title{ Lydia Pinkham Annual Advertising and Sales data}
\description{
  Annual domestic advertising and sales of Lydia E. Pinkham Medicine Company in thousands of dollars 1907-1960
}
\usage{data(pinkham)}
\format{
  A data frame with 54 observations on the 2 variables.
}
\source{
 Pankratz, A. (1991) Forecasting With Dynamic Regression Models, Wiley.  \cr\cr
}

\references{
  Wei, W. (1994) Time series analysis -  univariate and multivariate methods  
}

\keyword{datasets}
