% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNMR.R
\name{plotNMR}
\alias{plotNMR}
\title{Plot stacked 1H-NMR spectra}
\usage{
plotNMR(X, f1 = max(X[, 1]), f2 = min(X[, 1]), ppm = TRUE, title = "")
}
\arguments{
\item{X}{A matrix of spectra containing ppm values in the first column
and intensity values for each sample in the remaining columns. If a
matrix with samples x ppm is provided instead, it will be transposed
prior to plotting.}

\item{f1}{Chemical shift of the leftmost ppm value to plot.}

\item{f2}{Chemical shift of the rightmost ppm value to plot.}

\item{ppm}{Logical value. If \code{TRUE} the ppm values will be
as the plot x scale (default option). The variable indexes will used if
set to \code{FALSE}.}

\item{title}{Title of the plot (optional).}
}
\value{
A stacked plot of the full spectra or spectra region.
}
\description{
R function for plotting  1H-NMR spectra.
}
\examples{
data("urine")
plotNMR(urine)

}
\author{
Goncalo Graca, Imperial College London
}
