% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmethods.R
\name{print.margeno}
\alias{print.margeno}
\title{Print Method for margeno Objects}
\usage{
\method{print}{margeno}(x, ...)
}
\arguments{
\item{x}{An object of class "margeno" containing genetic data}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns NULL
}
\description{
Displays a summary of genetic data including sample counts, genomic sites,
and a preview of the genotype matrix
}
\examples{
print(gm1001g$geno)
}
\seealso{
\code{\link[=margeno]{margeno()}}
}
