% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1DNMR.R
\name{read1DNMR}
\alias{read1DNMR}
\title{Import 1d 1H-NMR spectra}
\usage{
read1DNMR(filepaths, xmax, xmin, npoints = 64000, simplify = TRUE)
}
\arguments{
\item{filepaths}{Full path to the processed Bruker data folder
\code{/pdata/}.}

\item{xmax}{Maximum value of the chemical shift scale for interpolation.}

\item{xmin}{Minimum value of the chemical shift scale for interpolation.}

\item{npoints}{Number of points to interpolate spectral intensities.}

\item{simplify}{If only the ppm and spectral intensities are to be
outputted. The default is \code{TRUE}. If set to \code{FALSE}, a list
list with ppm, intensity matrix and spectra names obtained from the title
file from the \code{/pdata/} folder will be outputted.}
}
\value{
The end result is outputted as a list with ppm, intensity matrix
and spectra names from title. If \code{simplify} is set to \code{TRUE},
a matrix with the ppm scale and spectra intensities will be returned.
}
\description{
Functions to import 1d spectra from Bruker format,
interpolate to get same chemical shift scale and export as data matrix.
}
\examples{
s <- read1DNMR(spaths, xmax=18, xmin=-1, npoints = 64000, simplify=TRUE)
}
\author{
Goncalo Graca, Imperial College London
}
