% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fjolst_en.R
\name{read_stations}
\alias{read_stations}
\title{read_stations}
\usage{
read_stations(year = NULL, gid = NULL, month = NULL, sq = NULL,
  cruise = NULL, id = NULL)
}
\arguments{
\item{year}{XXX}

\item{gid}{gear id}

\item{month}{XXX}

\item{sq}{statistical rectangle}

\item{cruise}{XXX}

\item{id}{XXX}
}
\description{
XXX
}

