% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveIntegration.R
\name{saveIntegration}
\alias{saveIntegration}
\title{Save the results of the integration of 1H-NMR spectra}
\usage{
saveIntegration(resultObject, DirPath = "")
}
\arguments{
\item{resultObject}{A list containing containing several objects:
the spectra matrix \code{spectra}, data frame containing the integration
regions, the number of protons and chemical shift of each metabolite signal,
\code{integrationRegions}, the integration results (\code{integrals} and
other the metadata such as \code{baseline}, the groups labels \code{groups},
sample names \code{spNames}, and project name \code{projectName}.}

\item{DirPath}{Path to the user-defined results folder.}
}
\value{
An overlay of the integrated spectral region and boxplots are saved if
sample group information is provided in the \code{groups} parameter or
otherwise barplots of each sample integral are saved.
Images and integration results table are saved to the user-defined folder
\code{DirPath}.
}
\description{
R function for plotting and exporting  1H-NMR spectra integration results.
Plots are generated for each integrated peak and the integration result
of each sample or groups of samples (if a group label is provided).
Plots are saved as \code{.jpeg} and the integrals table as \code{.csv}.
}
\examples{
userDir <- tempdir()
saveIntegration(result, DirPath=userDir)
}
\author{
Goncalo Graca, Imperial College London
}
