% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARsadsfs.R
\name{sfs}
\alias{sfs}
\title{Calculate Site Frequency Spectrum}
\usage{
sfs(gm, folded = FALSE, nozero = TRUE)
}
\arguments{
\item{folded}{Logical, whether to fold the spectrum. Default TRUE}

\item{nozero}{Logical, whether to remove zero counts. Default TRUE}

\item{AC}{Vector of allele counts}

\item{N}{Number of samples}

\item{ploidy}{Ploidy level of the organism}
}
\value{
An object of class "sfs" containing the site frequency spectrum
}
\description{
Calculate Site Frequency Spectrum
}
\examples{
# Calculate SFS from allele counts
allele_counts <- c(1,1,0,2,0,1,1,0,0,2,30)
sfs_result <- sfs(allele_counts, N=50, ploidy=2)
}
