% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shy1d.R
\name{shy1d}
\alias{shy1d}
\title{1D SHY using 1H-NMR spectra and other Y data}
\usage{
shy1d(X, Y, YvarName = "", ppm = NULL, expansion = NULL, title = "")
}
\arguments{
\item{X}{A matrix of spectra containing ppm values in the first row
and intensity values for each sample in the remaining rows If a
matrix with ppm x samples is provided instead, it will be transposed
prior to calculations and plotting.}

\item{Y}{A matrix of samples as rows and columns as variables, e.g.
a processed LC-MS output or or a low-resolution spectroscopy data
matrix (e.g. infrared, Raman or UV-VIS spectra).}

\item{YvarName}{Name of the variable to correlate with the NMR spectra as it
is represented in the Y matrix.}

\item{expansion}{A vector containing the end and start of a spectral region
in ppm. The default value is \code{NULL}, which will plot the full chemical
shift range.}
}
\value{
A a plot of the chemical shift vs covariance between the selected Y
variable and the NMR spectra. The correlation between the Y selected variable
and the NMR matrix will be color-coded on the covariance line.
}
\description{
Function to run 1D Statistical Hetero Spectroscopy (SHY) using
using 1H-NMR spectra as X matrix and other matrix-like Y dataset,
for instance an XCMS output for an LC-MS run or a low-resolution
spectroscopy data (e.g. infrared or Raman spectra).
}
\examples{
data("mtbls1")
X <- mtbls1[-1,]
Y <- mtbls1[-1,]
colnames(Y) <- mtbls1[1,]
shy1d(X, Y, YvarName="3.04955836737954")

}
\author{
Goncalo Graca, Imperial College London
}
