% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiskar.R
\name{skala_med_toldum}
\alias{skala_med_toldum}
\title{Skala með töldum}
\usage{
skala_med_toldum(lengdir)
}
\arguments{
\item{lengdir}{fyrirspurn á fiskar.lengdir}
}
\value{
fyrirspurn með sköluðum fjölda í lengdarbili
}
\description{
Þetta fall skala lengdardreifingar með töldum fiskum úr ralli
}

