\name{sparrows}
\alias{sparrows}
\docType{data}
\title{ Body measurements of sparrows}
\description{
Body measurements of 48 female sparrows.
}
\usage{data(sparrows)}
\format{
  A data frame with 48 observations on 5 variables
}

\source{
Manly, B. F. J. (1994). Multivariate Statistical Methods, second edition, Chapman and Hall.}


\keyword{datasets}
