% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk.R
\name{stk_mobile_icelandic}
\alias{stk_mobile_icelandic}
\title{stk_mobile_icelandic}
\usage{
stk_mobile_icelandic(con, correct = FALSE, vidmatch = FALSE, classify = FALSE)
}
\arguments{
\item{con}{ROracle connection}

\item{correct}{Correct mid}

\item{vidmatch}{Match vessels}

\item{classify}{Classify the mobileid}
}
\description{
stk_mobile_icelandic
}
