% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk.R
\name{stk_trail}
\alias{stk_trail}
\title{Lesa vms gögn}
\usage{
stk_trail(con, year)
}
\arguments{
\item{con}{XXX1}

\item{year}{XXX2}
}
\description{
Athugið aðgangur er takmarkaður við fáa notendur á Hafrannsóknastofnun
}
