% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy1d.R
\name{stocsy1d}
\alias{stocsy1d}
\title{1D STOCSY for 1H-NMR spectra}
\usage{
stocsy1d(spectra, driver, expansion = NULL)
}
\arguments{
\item{spectra}{A matrix of spectra containing ppm values in the first row
and intensity values for each sample in the remaining rows If a
matrix with ppm x samples is provided instead, it will be transposed
prior to calculations and plotting.}

\item{driver}{Chemical shift of the peak to which all other intensities
will be correlated. Covariance is also calculated between the driver peak
and all other points in the spectrum.}

\item{expansion}{A vector containing the end and start of a spectral region
in ppm. The default value is \code{NULL}, which will plot the full chemical
shift range.}
}
\value{
A a plot of 1d STOCSY, i.e. chemical shift vs covariance,
color-coded by the correlation coefficient.
}
\description{
Function to run 1D Statistical total correlation spectroscopy (STOCSY).
}
\examples{
data("mtbls1")
stocsy1d(mtbls1, 3.04)

}
\author{
Goncalo Graca, Imperial College London
}
