% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_fisktegundir}
\alias{veidibok_fisktegundir}
\title{Fisktegundir í veiðibókargrunni}
\usage{
veidibok_fisktegundir(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Id og nöfn fyrir fisktegundir í veiðibókargrunni
}
