% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_landsvaedi}
\alias{veidibok_landsvaedi}
\title{Landssvæði í veiðibókargrunni}
\usage{
veidibok_landsvaedi(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Landssvæði í veiðibókargrunni
}
