% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_notendur_svaedi}
\alias{veidibok_notendur_svaedi}
\title{Notendur í veiðibókargrunni}
\usage{
veidibok_notendur_svaedi(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Notendur í veiðibókargrunni.
}
