% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_thyngdarbil}
\alias{veidibok_thyngdarbil}
\title{Þyngdarbil veiðibókargrunnur}
\usage{
veidibok_thyngdarbil(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Hjálpartafla fyrir úrvinnslu
}
