% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_thyngdarstudlar}
\alias{veidibok_thyngdarstudlar}
\title{Þyngdarstuðlar veiðibókargrunnur}
\usage{
veidibok_thyngdarstudlar(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Meðalsamband lengdar og þyngdar hjá íslenskum laxi líklega $W = 0.0000215 L^2.83307$
}
