% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_vatnsfall}
\alias{veidibok_vatnsfall}
\title{Vatnsföll í veiðibókargrunni}
\usage{
veidibok_vatnsfall(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Upplýsingar um vatnsföll í veiðibókargrunni
}
