% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_veidarfaeri}
\alias{veidibok_veidarfaeri}
\title{Veiðarfæri í veiðibókargrunni}
\usage{
veidibok_veidarfaeri(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Veiðarfæri og agn í veiðibókargrunni. Fluga, Maðkur, Annað, Net eða Óákveðið
}
