% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_veidi}
\alias{veidibok_veidi}
\title{Veiði í veiðibókargrunni}
\usage{
veidibok_veidi(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Grunntafla fyrir veiði í veiðibókargrunni. Allar upplýsingar eru í \link{veidibok_veidibok} falli
}
