% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_veidibok}
\alias{veidibok_veidibok}
\title{Lax- og silungsveiðiskráningar veiðibókargrunnur}
\usage{
veidibok_veidibok(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Veitir aðgang að lax- og silungsveiðiskráningum. Einstaklingsskráningar með öllum helstu
upplýsingum úr öðrum töflum. Sameinað í Oracle-grunni.
}
