% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_veidistadur}
\alias{veidibok_veidistadur}
\title{Veiðistaðir í veiðibókargrunni}
\usage{
veidibok_veidistadur(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Upplýsingar um veiðistaði í veiðibókargruni
}
