% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veidibok.R
\name{veidibok_veidisvaedi}
\alias{veidibok_veidisvaedi}
\title{Veiðisvæði í veiðibókargrunni}
\usage{
veidibok_veidisvaedi(con)
}
\arguments{
\item{con}{src_oracle tenging við oracle}
}
\description{
Yfirlit yfir veiðisvæði í veiðibókargrunni.
}
