% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms.R
\name{vms}
\alias{vms}
\title{VMS data}
\arguments{
\item{con}{connection to mar database}

\item{t1}{Start date in the format "YYYY-MM-DD"}

\item{t2}{End date in the format "YYYY-MM-DD"}
}
\description{
Athugið aðgangur er takmarkaður við fáa notendur á Hafrannsóknastofnun
}
