\name{waves}
\alias{waves}
\docType{data}
\title{Time series of ocean wave height measurements}
\description{Ocean wave height measurements from an wire wave gauge and an infrared wave gauge}

\usage{data(waves)}
\format{
  A data frame with 4096 observations on the following 2 variables.
  \describe{
  \item{wire.gauge}{ height of ocean waves from wire wave gauge }
  \item{ir.gauge}{ height of ocean waves from infrared wave gauge}
  }
}
\details{
Time series of  ocean wave height measurements (sampling = 1/ 30 seconds)
}
\source{
 Applied Physics Laboratory (Andy Jessup)
}
\references{

Jessup, A. T., Melville, W. K., Keller, W. C. (1991). Breaking Waves Affecting Microwave Backscatter: Detection and Verification (1991). Journal of Geophysical Research, 96, C11, 20,547--59.\cr\cr

Percival, D. B. (1993). Spectral Analysis of Univariate and Bivariate Time Series, Chapter 11 of "Statistical Methods for Physical Science," Stanford, J. L.  and Vardeman, S. B. (Eds), Academic Press\cr\cr


}

\keyword{datasets}
