% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata.R
\name{ata_hafsyni}
\alias{ata_hafsyni}
\alias{ata_thurrvigt_tegund}
\alias{ata_rummal_thattur}
\alias{ata_thurrvigt_thattur}
\alias{ata_talningamadur}
\alias{ata_astand}
\alias{ata_votvigt}
\alias{ata_verkefni}
\alias{ata_thurrvigt}
\alias{ata_tegundir_hopar}
\alias{ata_talning}
\alias{ata_stod}
\alias{ata_stig}
\alias{ata_sofnunartaeki}
\alias{ata_snid}
\alias{ata_shortcut}
\alias{ata_rummal}
\alias{ata_prosomlengd}
\alias{ata_lengdir}
\alias{ata_leidangur}
\alias{ata_kynthroski}
\alias{ata_hopar_stig}
\alias{ata_hopar}
\alias{ata_flurljomun}
\alias{ata_eggjaframleidsla}
\title{Ástusýni}
\usage{
ata_hafsyni(con)

ata_thurrvigt_tegund(con)

ata_rummal_thattur(con)

ata_thurrvigt_thattur(con)

ata_talningamadur(con)

ata_astand(con)

ata_votvigt(con)

ata_verkefni(con)

ata_thurrvigt(con)

ata_tegundir_hopar(con)

ata_talning(con)

ata_stod(con)

ata_stig(con)

ata_sofnunartaeki(con)

ata_snid(con)

ata_shortcut(con)

ata_rummal(con)

ata_prosomlengd(con)

ata_lengdir(con)

ata_leidangur(con)

ata_kynthroski(con)

ata_hopar_stig(con)

ata_hopar(con)

ata_flurljomun(con)

ata_eggjaframleidsla(con)
}
\arguments{
\item{con}{tenging við mar}
}
\description{
Ástusýni
}
