% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_mar.R
\name{cn_mar}
\alias{cn_mar}
\title{Conform to mar column heading}
\usage{
cn_mar(con)
}
\arguments{
\item{con}{Tenging við Oracle}
}
\value{
SQL fyrirspurn
}
\description{
The function converts column of a dataframe, returning "_"
where there is a ".".
}
