% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARsadsfs.R
\name{expsfs}
\alias{expsfs}
\title{Generate Expected Site Frequency Spectrum}
\usage{
expsfs(gm, folded = FALSE, nozero = TRUE)
}
\arguments{
\item{folded}{Logical, whether to fold the spectrum. Default TRUE}

\item{nozero}{Logical, whether to remove zero counts. Default TRUE}

\item{lenAC}{Length of allele count vector. In other words, total number of SNPs surveyed.}

\item{N}{Number of samples}

\item{ploidy}{Ploidy level of the organism}
}
\value{
An object of class "sfs" containing the expected site frequency spectrum
}
\description{
Generate Expected Site Frequency Spectrum
}
\examples{
# Generate expected SFS
exp_sfs <- expsfs(lenAC=1000, N=100, ploidy=2)
}
