% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magi.R
\name{faeda}
\alias{faeda}
\alias{faeda_tegundir}
\alias{faeda_ranfiskar}
\alias{faeda_thyngdir}
\alias{faeda_lengdir}
\title{Magasyni Hafrannsóknarstofnunar}
\usage{
faeda_tegundir(mar)

faeda_ranfiskar(mar)

faeda_thyngdir(mar)

faeda_lengdir(mar)
}
\arguments{
\item{mar}{tenging við oraclegrunn}
}
\value{
fyrirspurn
}
\description{
Magasyni Hafrannsóknarstofnunar
}
