% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomaps.R
\name{genomaps}
\alias{genomaps}
\title{Create a genomaps object}
\usage{
genomaps(geno, maps)
}
\arguments{
\item{geno}{A margeno object.}

\item{maps}{A marmaps object.}
}
\value{
A genomaps object, a list of the \code{geno} and \code{maps} inputs.
}
\description{
The genomaps object is the central data structure for the mar package. It combines the genetic data
from the margeno object and the spatial information from the marmaps object, providing a
comprehensive representation of the genetic and geographic data, connected by the sample IDs.
}
\examples{
# gm1001g is itself a genomaps object, rebuild it from its two components
geno <- gm1001g$geno
maps <- gm1001g$maps
gm <- genomaps(geno, maps)

all.equal(gm, gm1001g)
\dontrun{
# genomaps() will fail if sample names are not matching
maps_bad <- gm1001g$maps
maps_bad$sample.id <- 1:1004
gm <- genomaps(geno, maps_bad)
}
}
\seealso{
\code{\link[=margeno]{margeno()}} and \code{\link[=marmaps]{marmaps()}} for the two components, and
\code{\link[=MARsampling]{MARsampling()}}, \code{\link[=MARextinction]{MARextinction()}} and \code{\link[=MARtheory]{MARtheory()}} for using genomaps.
}
