% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gm1001g}
\alias{gm1001g}
\title{gm1001g}
\format{
\subsection{\code{gm1001g}}{

An S3 \code{genomaps} object created using \code{\link[=genomaps]{genomaps()}} function.
\describe{
\item{geno}{An S3 \code{margeno} object. See \code{\link[=margeno]{margeno()}} function.}
\item{maps}{An S3 \code{marmaps} object. See \code{\link[=marmaps]{marmaps()}} function.}
}
}
}
\source{
\url{https://1001genomes.org/data/GMI-MPI/releases/v3.1/SNP_matrix_imputed_hdf5/}
}
\usage{
gm1001g
}
\description{
Example genomaps object created from Arabidopsis 1001G project
}
\keyword{datasets}
