% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiskar.R
\name{lesa_kvarnir}
\alias{lesa_kvarnir}
\title{Kvarnir}
\usage{
lesa_kvarnir(mar)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}
}
\value{
dataframe
}
\description{
Fallid myndar tengingu við toflu kvarnir
í fiskar gagnagrunninum.
}
\examples{
mar <- dplyrOracle::src_oracle("mar")
dplyr::glimpse(lesa_numer(mar))
}

