% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fjolst.R
\name{lesa_stodvar}
\alias{lesa_stodvar}
\title{lesa_stodvar}
\usage{
lesa_stodvar(ar = NULL, veidarfaeri = NULL, man = NULL, reitur = NULL,
  leidangur = NULL, synis.id = NULL)
}
\arguments{
\item{ar}{XXX}

\item{veidarfaeri}{XXX}

\item{man}{XXX}

\item{reitur}{XXX}

\item{leidangur}{XXX}

\item{synis.id}{XXX}
}
\description{
XXX
}

