% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar-package.R
\docType{package}
\name{mar}
\alias{mar-package}
\alias{mar}
\title{mar: Fitting Mutations-Area Relationships using population
genomic samples of a species}
\description{
mar package provides functions to approximate the genetic diversity
loss in a species from habitat reduction.
}
\section{Key functions}{

\describe{
\item{\code{MARsampling()}}{samples genetic diversity across varying spatial scales}
\item{\code{MARcalc()}}{fit mutation area relation model}
\item{\code{MARextinction()}}{projects the expected loss of genetic diversity under hypothetical area reduction scenarios}

}
}

\seealso{
Useful links:
\itemize{
\item{Package repository: \url{https://github.com/meixilin/mar.git}}
\item{Package paper: \url{https://www.biorxiv.org/content/10.1101/2025.09.09.675155v1.full}}
}
}
\keyword{internal}
