% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_info.R
\name{mar_fields}
\alias{mar_fields}
\title{mar_fields}
\usage{
mar_fields(mar, table)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}

\item{table}{character vector specifying table names, e.g. "fiskar.stodvar"}
}
\value{
dataframe
}
\description{
Fallið ...
}
\note{
NB - ætti frekar að laga dplyrOracle:::db_query_fields.OraConnection??
}
