% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mar_info.R
\name{mar_tables}
\alias{mar_tables}
\title{mar_tables}
\usage{
mar_tables(mar, schema)
}
\arguments{
\item{mar}{src_oracle tenging við oracle}

\item{schema}{character vector specifying table names, e.g. "fiskar". If missing
information from all tables are returned.}
}
\value{
dataframe
}
\description{
Fallið myndar tengingu við schema all_tables
}

