% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARcalc.R
\name{MARcalc}
\alias{MARcalc}
\title{Calculate the mutations-area relationship (MAR) using the power-law model}
\usage{
MARcalc(mardf, Mtype = .Mtype, Atype = .Atype)
}
\arguments{
\item{mardf}{A data frame with columns of diversity metric and area.
Output from \link{MARsampling}, \link{MARextinction} or \link{MARtheory}.}

\item{Mtype}{The diversity metric to use. Default is "M" (number of mutation), allowed values are M, E, thetaw, thetapi.}

\item{Atype}{The area metric to use. Default is "A" (area), allowed values are A, Asq, N.}
}
\value{
A fitted power model object from \code{\link[sars:sar_power]{sars::sar_power()}}
}
\description{
This function calculates the MAR using the power-law model from the provided data frame.
The data frame should contain columns for the diversity metric (e.g. number of mutation, nucleotide diversity)
and the area metric (e.g. occupied area, total squared area).
}
\examples{
mardf <- MARsampling(gm1001g, nrep = 2, xfrac = 0.1, myseed = 42)
mar <- MARcalc(mardf, Mtype = "M", Atype = "A")
summary(mar)

# the same power law model fitted against sample size rather than area
MARcalc(mardf, Mtype = "M", Atype = "N")
}
\seealso{
\code{\link[=MARsampling]{MARsampling()}}, \code{\link[=MARextinction]{MARextinction()}} and \code{\link[=MARtheory]{MARtheory()}} for generating the input \code{mardf}.
}
