% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARextinction.R
\name{MARextinction}
\alias{MARextinction}
\title{Simulate the mutations-area relationship (MAR) extinction process}
\usage{
MARextinction(
  gm,
  scheme = .MARsampling_schemes,
  nrep = 10,
  xfrac = 0.01,
  animate = FALSE,
  myseed = NULL
)
}
\arguments{
\item{gm}{A genomaps object.}

\item{scheme}{The sampling scheme to use for the extinction process. Default is "random", allowed values are random, inwards, outwards, southnorth, northsouth.}

\item{nrep}{The number of extinction replicates to perform. Default is 10.}

\item{xfrac}{The fraction of cells to be randomly removed at each extinction step. Default is 0.01 or one raster cell if there are less than 100 cells.}

\item{animate}{If TRUE, the function will animate the extinction process. Default is FALSE.}

\item{myseed}{An optional seed value to ensure reproducibility. Implemented as \code{set.seed(myseed)}. Default is NULL.}
}
\value{
A data frame containing the area and diversity metrics for each step of the extinction process.
}
\description{
This function performs a simulated extinction process on a map of genomic samples, similar to the
MARsampling function, but with the extinction happening at the cell level rather than
circling the grid with boxes. The function returns a data frame containing the area and
diversity metrics for each step of the extinction process.
}
\examples{
\dontrun{
extdf <- MARextinction(gm1001g)
}

}
