% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomaps.R
\name{margeno}
\alias{margeno}
\title{Create a margeno object}
\usage{
margeno(sample.id, variant.id, position, chromosome, genotype, ploidy)
}
\arguments{
\item{sample.id}{Character or integer or numeric vector of unique sample IDs}

\item{variant.id}{Integer vector of unique variant IDs}

\item{position}{Integer vector of variant positions. Or NULL.}

\item{chromosome}{Character or integer vector of chromosome IDs. Or NULL.}

\item{genotype}{Matrix of genotypes, where rows represent samples and columns represent variants. Each value represents the number of alternative alleles.}

\item{ploidy}{Numeric value of ploidy. \emph{Warning} ploidy other than 2 can be used, but result interpretation is not guaranteed.}
}
\value{
A margeno object
}
\description{
Create a margeno object
}
\examples{
sample_id <- c("sample1", "sample2", "sample3")
variant_id <- 1:4
position <- as.integer(c(100, 200, 300, 400)) # position has to be integer
chromosome <- c("chr1", "chr1", "chr1", "chr2")
position <- NULL
chromosome <- NULL # if no position or chromosome
genotype <- matrix(c(1, 2, 0, 0, 1, 1, 2, 0, 1, 2, 2, 1), nrow = 4, ncol = 3)
ploidy <- 2
margeno(sample_id, variant_id, position, chromosome, genotype, ploidy)
}
