% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARsampling.R
\name{MARsampling}
\alias{MARsampling}
\title{MAR sampling wrapper function}
\usage{
MARsampling(
  gm,
  scheme = .MARsampling_schemes,
  nrep = 10,
  quorum = FALSE,
  animate = FALSE,
  myseed = NULL
)
}
\arguments{
\item{gm}{a [genemaps] object created by [genemaps()]}

\item{scheme}{sampling schemes for spatial data. allowed are \code{\link{.MARsampling_schemes}}.}

\item{nrep}{number of replicates.}

\item{quorum}{require all sampling grid to have samples. default is FALSE.}

\item{animate}{play an animation of the sampling boxes. default is FALSE.}

\item{myseed}{set seed for reproducibility. default is NULL.}
}
\value{
a [marsamp] object. consist of a data frame.
}
\description{
MAR sampling wrapper function
}
