% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARsampling.R
\name{MARsampling}
\alias{MARsampling}
\title{MAR sampling wrapper function}
\usage{
MARsampling(
  gm,
  scheme = .MARsampling_schemes,
  nrep = 10,
  xfrac = 0.01,
  quorum = TRUE,
  animate = FALSE,
  myseed = NULL
)
}
\arguments{
\item{gm}{a \link{genomaps} object created by \code{\link[=genomaps]{genomaps()}}}

\item{scheme}{sampling schemes for spatial data. allowed are random, inwards, outwards, southnorth, northsouth.}

\item{nrep}{number of replicates.}

\item{xfrac}{fraction of the range to use for the step size in the sampling scheme.}

\item{quorum}{require all sampling grid to have samples. default is FALSE.}

\item{animate}{play an animation of the sampling boxes. default is FALSE.}

\item{myseed}{set seed for reproducibility. default is NULL.}
}
\value{
a \code{marsamp} object. consist of a data frame.
}
\description{
MAR sampling wrapper function
}
