% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiQuant.R
\name{multiQuant}
\alias{multiQuant}
\title{Integration and quantification of 1H-NMR spectra}
\usage{
multiQuant(
  M,
  reg,
  reference = "ERETIC",
  refConc = 10,
  SpNames = NULL,
  baseline = TRUE,
  groups = NA,
  projectName = ""
)
}
\arguments{
\item{M}{Matrix of 1H-NMR spectra where samples are arranged in rows
and chemical shift values (ppm) in columns. The matrix will be transposed
otherwise. Sample names should not be included in this matrix, only ppm
values.}

\item{reg}{Data frame containing as columns: \code{metabolite}, \code{ppm},
number of protons corresponding to the signal \code{nprotons}, start and
end of the signal in ppm (\code{ppm.start} and \code{ppm.end}, respectively)}

\item{reference}{Exact name of the reference metabolite or compound, as
named in the \code{reg} data frame.}

\item{refConc}{The concentration of the reference compound/metabolite in mM}

\item{SpNames}{A string containing the sample names. The default
is \code{NULL}.}

\item{baseline}{A logical value to specify if baseline should be performed.
The default value is \code{TRUE}.}

\item{groups}{A string of values or characters that specify any sample
groups (e.g. controls and disease). The default is \code{NA}.}

\item{projectName}{The name of the project, which will be appended to the
results files upon saving.}
}
\value{
A list containing containing several objects:
the spectra matrix \code{spectra}, data frame containing the integration
regions, the number of protons and chemical shift of each metabolite signal,
\code{integrationRegions}, the integration and quantification results
(\code{integrals} and \code{quantification}) and other the metadata such as
\code{baseline}, reference name(\code{reference}) and concentration
(\code{refConc}), the groups labels \code{groups}, sample names
\code{spNames}, and project name \code{projectName}.
}
\description{
R function for baseline correction, integration and quantification
The \code{ptw} package is required to perform baseline correction.
The main function performs integration on multiple spectra regions with or
without baseline correction and calculates the respective metabolite
concentration given a quantification reference (e.g. TSP or ERETIC signal).
}
\examples{
data("mtbls1")
reg <- data.frame(metabolite = c("Creatine", "Creatinine", "Glucose"), 
                  nprotons = c(3, 2, 1), ppm=c(3.04, 4.06, 5.23), 
                  ppm.start=c(3.046, 4.068, 5.26), 
                   ppm.end=c(3.040, 4.054, 5.23))
quanResult <- multiQuant(urine, reg, reference="Glucose", refConc=1, 
                         SpNames=NULL, baseline=TRUE, groups=NA, 
                         projectName="MTBLS1")
}
\author{
Goncalo Graca, Imperial College London
}
