% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutdiv.R
\name{mutdiv.cells}
\alias{mutdiv.cells}
\title{Calculate Genetic Diversity for Specified Cells}
\usage{
mutdiv.cells(gm, gmarea, cellids)
}
\arguments{
\item{gm}{A genomaps object containing genetic data and geographic informations, created by \code{\link[=genomaps]{genomaps()}} function.}

\item{gmarea}{Raster file that contains area size of each cell}

\item{cellids}{Vector of cell IDs to analyze}
}
\value{
A list containing:
\item{N}{Number of samples}
\item{M}{Number of segregating sites}
\item{E}{Number of endemic segregating sites}
\item{thetaw}{Watterson's theta estimate}
\item{thetapi}{Pi (pairwise) diversity estimate}
\item{A}{Total area with data}
\item{Asq}{Area of the cells being analyzed}
}
\description{
Calculates genetic diversity metrics for a specific set of cells. This function is particularly
useful for extinction simulations.
}
\examples{
# Calculate genetic diversity for a specific set of cells
gmarea <- mar:::.areaofraster(gm1001g$maps$samplemap)
cell_ids <- c(613, 726, 727)
div <- mutdiv.cells(gm1001g, gmarea, cellids = cell_ids)
}
