% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutdiv.R
\name{mutdiv.gridded}
\alias{mutdiv.gridded}
\title{Calculate Genetic Diversity in a Gridded Bounding Box}
\usage{
mutdiv.gridded(gm, gmarea, sm, bbox, revbbox = FALSE)
}
\arguments{
\item{gm}{A genomaps object containing genetic data and geographic informations, created by \code{\link[=genomaps]{genomaps()}} function.}

\item{gmarea}{Raster file that contains area size of each cell}

\item{bbox}{Numeric vector of length 4 specifying the bounding box coordinates c(r1, r2, c1, c2)}

\item{revbbox}{Logical, whether to reverse/invert the bounding box selection. Default FALSE}
}
\value{
A list containing:
\item{N}{Number of samples}
\item{M}{Number of segregating sites}
\item{E}{Number of endemic segregating sites}
\item{thetaw}{Watterson's theta estimate}
\item{thetapi}{Pi (pairwise) diversity estimate}
\item{A}{Total area with data}
\item{Asq}{Area of the bounding box square}
}
\description{
Calculate Genetic Diversity in a Gridded Bounding Box
}
\examples{
# Calculate mutation diversity in a 10x10 grid region
gmarea <- mar:::.areaofraster(gm1001g$maps$samplemap)
div <- mutdiv.gridded(gm1001g, gmarea, bbox = c(1, 10, 2, 8))
}
