% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.marextinct}
\alias{plot.marextinct}
\title{Plot Method for Mutations Extinction Curves}
\usage{
\method{plot}{marextinct}(x, z = NULL, Mtype = .Mtype, Atype = .Atype, ...)
}
\arguments{
\item{x}{A data frame containing extinction data}

\item{z}{Fitted extinction curve parameter}

\item{Mtype}{Type of genetic diversity metric to plot}

\item{Atype}{Type of area metric to plot}

\item{...}{Additional arguments passed to plot}
}
\value{
Invisibly returns NULL
}
\description{
Creates a plot showing relationship between area loss and genetic diversity loss
}
\examples{
\dontrun{
plot(marextinct_object, z = 0.25)
}
}
