% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.marmaps}
\alias{plot.marmaps}
\title{Plot Method for marmaps Class}
\usage{
\method{plot}{marmaps}(x, ...)
}
\arguments{
\item{x}{An object of class "marmaps" containing sample mapping information}

\item{...}{Additional arguments passed to plot}
}
\value{
Invisibly returns NULL
}
\description{
Creates a spatial plot of sample locations and density
}
\examples{
plot(gm1001g$maps)
}
