% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmethods.R
\name{plot.sfs}
\alias{plot.sfs}
\title{Plot Method for Site Frequency Spectrum}
\usage{
\method{plot}{sfs}(x, ...)
}
\arguments{
\item{x}{An object of class "sfs" containing site frequency spectrum data}

\item{...}{Additional arguments passed to barplot}
}
\value{
Invisibly returns NULL
}
\description{
Creates a barplot of the site frequency spectrum
}
\examples{
sfs_object <- sfs(AC = c(1, 1, 0, 2, 0, 1, 1, 0, 0, 2, 30), N = 50, ploidy = 2)
plot(sfs_object)
}
