% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveQuantification.R
\name{saveQuantification}
\alias{saveQuantification}
\title{Save the results of the quantification of 1H-NMR spectra}
\usage{
saveQuantification(resultObject, DirPath = "")
}
\arguments{
\item{resultObject}{A list containing containing several objects:
the spectra matrix \code{spectra}, data frame containing the integration
regions, the number of protons and chemical shift of each metabolite signal,
\code{integrationRegions}, the integration and quantification results
(\code{integrals} and \code{quantification}) and other the metadata such as
\code{baseline}, reference name(\code{reference}), reference concentration
(\code{refConc}), omit reference from results (\code{removeRef}),
the groups labels \code{groups}, sample names
\code{spNames}, and project name \code{projectName}.}

\item{DirPath}{Path to the user-defined results folder.}
}
\value{
An overlay of the integrated spectral region and boxplots are saved if
sample group information is provided in the \code{groups} parameter or
otherwise barplots of each sample concentration are saved.
Images and quantification results table are saved to the user-defined folder
\code{DirPath}.
}
\description{
R function for plotting and exporting  1H-NMR spectra quantification results.
Plots are generated for each integrated peak and the quantification result
of each sample or groups of samples (if a group label is provided).
Plots are saved as \code{.jpeg} and the quantification table as \code{.csv}.
}
\examples{
data("mtbls1")
reg <- data.frame(metabolite = c("Creatine", "Creatinine", "Glucose"), 
                  nprotons = c(3, 2, 1), ppm=c(3.04, 4.06, 5.23), 
                  ppm.start=c(3.046, 4.068, 5.26), 
                   ppm.end=c(3.040, 4.054, 5.23))
quanResult <- multiQuant(urine, reg, reference="Glucose", refConc=1, 
                         SpNames=NULL, baseline=TRUE, groups=NA, 
                         projectName="MTBLS1")
userDir <- tempdir()
saveQuantification(quanResult, DirPath=userDir)

}
\author{
Goncalo Graca, Imperial College London
}
