% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparser.R
\name{vcf_parser}
\alias{vcf_parser}
\title{Convert VCF file to GDS format}
\usage{
vcf_parser(vcf.fn, ploidy = 2)
}
\arguments{
\item{vcf.fn}{Path to the input VCF file. Can be either .vcf or .vcf.gz format}

\item{gds.fn}{Optional. Path for the output GDS file. If NULL, will use the same name as vcf file
with .gds extension}

\item{opengds}{Logical. If TRUE, opens and returns the GDS file handle. If FALSE, returns the
path to the created GDS file. Default is FALSE}
}
\value{
If opengds=TRUE, returns an opened GDS file connection. If opengds=FALSE, returns the
path to the created GDS file as a character string
}
\description{
This function converts a VCF (Variant Call Format) file to GDS (Genomic Data Structure) format
using SeqArray package.
}
\examples{
\dontrun{
# Convert VCF to GDS
gds_file <- vcf_parser("input.vcf")

# Convert and open GDS file
gds_conn <- vcf_parser("input.vcf.gz", opengds = TRUE)
}
}
